experiment with programmatic endpoints
