+++ components/subscription.yaml
@@ -1,11 +1,11 @@
-apiVersion: dapr.io/v1alpha1
-kind: Subscription
-metadata:
-  name: weather-forecast-pub-sub
-spec:
-  topic: weather-forecasts
-  route: /SendWeatherForecast
-  pubsubname: weather-forecast-pub-sub
-scopes:
-- node-app
-- dotnet-app
+# apiVersion: dapr.io/v1alpha1
+# kind: Subscription
+# metadata:
+#   name: weather-forecast-pub-sub
+# spec:
+#   topic: weather-forecasts
+#   route: /SendWeatherForecast
+#   pubsubname: weather-forecast-pub-sub
+# scopes:
+# - node-app
+# - dotnet-app
